requirements.txt
pillow
easyocr
pandas
rapidfuzz
pywin32
pypiwin32

